refactor: book sitemap 병합 및 route 변경
