+++ src/app/sitemap.ts
@@ -1,17 +1,13 @@
 import type { MetadataRoute } from 'next';
 
-import { default as booksSitemap } from './book/[bookId]/sitemap';
-import { default as bookshelvesSitemap } from './bookshelf/[bookshelfId]/sitemap';
-import { default as bookGroupSitemap } from './group/[groupId]/sitemap';
+import { default as bookSitemap } from './book/sitemap';
+import { default as bookshelfSitemap } from './bookshelf/sitemap';
+import { default as bookGroupSitemap } from './group/sitemap';
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   return [
     {
       url: `${process.env.NEXT_HOST}/bookarchive`,
-      lastModified: new Date(),
-    },
-    {
-      url: `${process.env.NEXT_HOST}/book/search`,
       lastModified: new Date(),
     },
     {
@@ -22,8 +18,8 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       url: `${process.env.NEXT_HOST}/profile/me`,
       lastModified: new Date(),
     },
-    ...(await booksSitemap()),
-    ...(await bookshelvesSitemap()),
+    ...(await bookSitemap()),
+    ...(await bookshelfSitemap()),
     ...(await bookGroupSitemap()),
   ];
 }
